PP-6771 Log filters for transaction search and download
Log the filters that are applied to transaction searches and CSV
downloads as well as the gateway account ids.

Additionally log a `multiple_accounts` field to quickly get whether
the request was for all gateway accounts for a user.

The intention of this is to be able to easily group requests by their
filters in Splunk so we can see which combinations of filters cause the
search to take longer.

Extract logging when csv streaming is complete to a function in
transaction_service to reduce duplication.
